requirements.txt
blinker==1.9.0
click==8.2.1
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fpdf2==2.8.3
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pillow==11.3.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
gunicorn==23.0.0
